+++ photutils/aperture/rectangle.py
@@ -93,17 +93,17 @@ class RectangularMaskMixin:
         masks = []
         for bbox, edges in zip(self._bbox, self._centered_edges, strict=True):
             ny, nx = bbox.shape
+            theta_rad = self.theta.to(u.radian).value
             mask = rectangular_overlap_grid(edges[0], edges[1], edges[2],
                                             edges[3], nx, ny, w, h,
-                                            self._theta_radians, 0, subpixels)
+                                            theta_rad, 0, subpixels)
 
             # subtract the inner circle for an annulus
             if hasattr(self, 'w_in'):
                 mask -= rectangular_overlap_grid(edges[0], edges[1], edges[2],
                                                  edges[3], nx, ny, self.w_in,
-                                                 self.h_in,
-                                                 self._theta_radians, 0,
-                                                 subpixels)
+                                                 self.h_in, theta_rad,
+                                                 0, subpixels)
 
             masks.append(ApertureMask(mask, bbox))
 
@@ -117,10 +117,11 @@ class RectangularMaskMixin:
         """
         Calculate half of the bounding box extents of an ellipse.
         """
+        theta_rad = theta.to(u.radian).value
         half_width = width / 2.0
         half_height = height / 2.0
-        sin_theta = math.sin(theta)
-        cos_theta = math.cos(theta)
+        sin_theta = math.sin(theta_rad)
+        cos_theta = math.cos(theta_rad)
         x_extent1 = abs((half_width * cos_theta) - (half_height * sin_theta))
         x_extent2 = abs((half_width * cos_theta) + (half_height * sin_theta))
         y_extent1 = abs((half_width * sin_theta) + (half_height * cos_theta))
@@ -138,10 +139,11 @@ class RectangularMaskMixin:
         Used for creating `~matplotlib.patches.Rectangle` patch for the
         aperture.
         """
+        theta_rad = theta.to(u.radian).value
         half_width = width / 2.0
         half_height = height / 2.0
-        sin_theta = math.sin(theta)
-        cos_theta = math.cos(theta)
+        sin_theta = math.sin(theta_rad)
+        cos_theta = math.cos(theta_rad)
         xshift = (half_height * sin_theta) - (half_width * cos_theta)
         yshift = -(half_height * cos_theta) - (half_width * sin_theta)
 
@@ -211,12 +213,11 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
         self.positions = positions
         self.w = w
         self.h = h
-        self._theta_radians = 0.0  # defined by theta setter
         self.theta = theta
 
     @property
     def _xy_extents(self):
-        return self._calc_extents(self.w, self.h, self._theta_radians)
+        return self._calc_extents(self.w, self.h, self.theta)
 
     @property
     def area(self):
@@ -252,11 +253,11 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
         xy_positions, patch_kwargs = self._define_patch_params(origin=origin,
                                                                **kwargs)
         xy_positions = self._lower_left_positions(xy_positions, self.w,
-                                                  self.h, self._theta_radians)
+                                                  self.h, self.theta)
 
-        theta_deg = self._theta_radians * 180.0 / np.pi
+        angle = self.theta.to(u.deg).value
         patches = [mpatches.Rectangle(xy_position, self.w, self.h,
-                                      angle=theta_deg, **patch_kwargs)
+                                      angle=angle, **patch_kwargs)
                    for xy_position in xy_positions]
 
         if self.isscalar:
@@ -384,12 +385,11 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
             raise ValueError('"h_out" must be greater than "h_in"')
         self.h_in = h_in
 
-        self._theta_radians = 0.0  # defined by theta setter
         self.theta = theta
 
     @property
     def _xy_extents(self):
-        return self._calc_extents(self.w_out, self.h_out, self._theta_radians)
+        return self._calc_extents(self.w_out, self.h_out, self.theta)
 
     @property
     def area(self):
@@ -426,20 +426,20 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
                                                                **kwargs)
         inner_xy_positions = self._lower_left_positions(xy_positions,
                                                         self.w_in, self.h_in,
-                                                        self._theta_radians)
+                                                        self.theta)
         outer_xy_positions = self._lower_left_positions(xy_positions,
                                                         self.w_out,
                                                         self.h_out,
-                                                        self._theta_radians)
+                                                        self.theta)
 
         patches = []
-        theta_deg = self._theta_radians * 180.0 / np.pi
+        angle = self.theta.to(u.deg).value
         for xy_in, xy_out in zip(inner_xy_positions, outer_xy_positions,
                                  strict=True):
             patch_inner = mpatches.Rectangle(xy_in, self.w_in, self.h_in,
-                                             angle=theta_deg)
+                                             angle=angle)
             patch_outer = mpatches.Rectangle(xy_out, self.w_out, self.h_out,
-                                             angle=theta_deg)
+                                             angle=angle)
             path = self._make_annulus_path(patch_inner, patch_outer)
             patches.append(mpatches.PathPatch(path, **patch_kwargs))